# myapp/web/swagger-ui/swagger.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "HH33 API Specification",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://localhost",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/books": {
            "get": {
                "tags": [
                    "Book"
                ],
                "summary": "Список книг",
                "description": "Получить список книг с возможностью фильтрации по названию/описанию и авторам",
                "operationId": "101458f1cdc8a823172815d0d500b484",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Поиск по названию или описанию книги",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "author[]",
                        "in": "query",
                        "description": "Фильтр по id авторов",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    },
                    {
                        "name": "language[]",
                        "in": "query",
                        "description": "Фильтр по языкам",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    {
                        "name": "genre",
                        "in": "query",
                        "description": "Поиск по жанру",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "pages_from",
                        "in": "query",
                        "description": "Число страниц ОТ (включительно)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "pages_to",
                        "in": "query",
                        "description": "Число страниц ДО (включительно)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список книг",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BookResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Book"
                ],
                "summary": "Создать новую книгу",
                "operationId": "3096bbcd2efa33183e5c6ed1cf9749a9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Book"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Кинига создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка запроса",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpException"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/books/{id}": {
            "put": {
                "tags": [
                    "Book"
                ],
                "summary": "Обновление объекта книги",
                "operationId": "e4536fd96f4d58f87e35b5388d7e9871",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id изменяеемого книги",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Book"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Книга обновлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "запись не найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpException"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка запроса",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpException"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Book"
                ],
                "summary": "Удалить книгу",
                "operationId": "37e3352905a1a7ddaea880b6e7302615",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id удаляемой книги",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Книга успешно удалена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "message"
                                    ],
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Removed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Книга не найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpException"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка при удалении",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "errors"
                                    ],
                                    "properties": {
                                        "errors": {
                                            "type": "string",
                                            "example": "Delete failed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "BookResponse": {
                "title": "BookResponse",
                "description": "Book Model without id field",
                "required": [
                    "title",
                    "author_id",
                    "pages",
                    "language",
                    "genre",
                    "description"
                ],
                "properties": {
                    "title": {
                        "description": "Book title",
                        "type": "string",
                        "maxLength": 65535
                    },
                    "author_id": {
                        "description": "Author of the book",
                        "type": "integer"
                    },
                    "pages": {
                        "description": "Number of pages in the book",
                        "type": "integer"
                    },
                    "language": {
                        "description": "Language of the book",
                        "type": "string",
                        "maxLength": 65535
                    },
                    "genre": {
                        "description": "Genre of the book",
                        "type": "string",
                        "maxLength": 65535
                    },
                    "description": {
                        "description": "Short summary of the book",
                        "type": "text"
                    }
                },
                "type": "object"
            },
            "HttpException": {
                "description": "HttpException",
                "properties": {
                    "name": {
                        "description": "Краткое название ошибки",
                        "type": "string"
                    },
                    "message": {
                        "description": "Сообщение об ошибке",
                        "type": "string"
                    },
                    "code": {
                        "description": "Код внутренней ошибки",
                        "type": "integer"
                    },
                    "status": {
                        "description": "HTTP-статус ответа",
                        "type": "integer"
                    },
                    "type": {
                        "description": "Тип исключения",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "description": "Ответ ошибки валидации",
                "properties": {
                    "errors": {
                        "description": "Список ошибок по каждому полю",
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
...
            "Book": {
                "title": "Book",
                "description": "Book Model",
                "required": [
                    "title",
                    "author_id",
                    "pages",
                    "language",
                    "genre",
                    "description"
                ],
                "properties": {
                    "id": {
                        "description": "Unique identifier of the book",
                        "type": "integer",
                        "readOnly": true
                    },
                    "title": {
                        "description": "Book title",
                        "type": "string",
                        "maxLength": 65535
                    },
                    "author_id": {
                        "description": "Author of the book",
                        "type": "integer"
                    },
                    "pages": {
                        "description": "Number of pages in the book",
                        "type": "integer"
                    },
                    "language": {
                        "description": "Language of the book",
                        "type": "string",
                        "maxLength": 65535
                    },
                    "genre": {
                        "description": "Genre of the book",
                        "type": "string",
                        "maxLength": 65535
                    },
                    "description": {
                        "description": "Short summary of the book",
                        "type": "text"
                    }
                },
                "type": "object"
            }
        }
    },
...
}